refactor: 타입 위치 이동
