+++ src/api/cafeteria/entity.ts
@@ -1,5 +1,5 @@
 import { APIResponse } from 'interfaces/APIResponse';
-import { Dining } from 'interfaces/Cafeteria';
+import { Dining } from 'static/cafeteria';
 
 export interface DiningResponseType extends APIResponse {
   [index: number]: Dining;
