--- src/interfaces/Lecture.ts
@@ -1,5 +1,0 @@
-import { Lecture } from 'api/timetable/entity';
-
-export interface TimetableInfoFromLocalStorage {
-  [key: string]: Lecture[];
-}
