+++ src/pages/Cafeteria/PCCafeteriaPage/components/PCMealImage/index.tsx
@@ -1,5 +1,5 @@
 import { cn } from '@bcsdlab/utils';
-import { Dining } from 'interfaces/Cafeteria';
+import { Dining } from 'static/cafeteria';
 import NoPhoto from 'assets/svg/no-photography-pc.svg';
 import NoMeals from 'assets/svg/no-meals-pc.svg';
 import styles from './PCMealImage.module.scss';
