+++ src/utils/zustand/myLectures.ts
@@ -1,7 +1,10 @@
 import { Lecture } from 'api/timetable/entity';
-import { TimetableInfoFromLocalStorage } from 'interfaces/Lecture';
 import { create } from 'zustand';
 import { useShallow } from 'zustand/react/shallow';
+
+interface TimetableInfoFromLocalStorage {
+  [key: string]: Lecture[];
+}
 
 const MY_LECTURES_KEY = 'my-lectures';
